[lvm_multi_vpg_same_pv] LogicalVolumeManager when trying to perform GetLogicalVolumes and has multiple VolumeGroup located on different PhysicalVolume that are named the same (e.g. pv0), the wrong physical volume is used since we search by name from all volumeGroup instead of by name of physicalvolume only from list of Volumegroup.
E.g.
Disk has 2 partitions:
/dev/sda2
/dev/sda3

Volumegroup (VG1) with single LogicalVolume (LV1) on /dev/sda2
So in the VG1 metadata it states:
 - pv0 located on /dev/sda2
 - LV1 has a single segment with single stripe on pv0

Volumegroup (VG2) with single LogicalVolume (LV2) on /dev/sda3
So in the VG2 metadata it states:
 - pv0 located on /dev/sda3
 - LV2 has a single segment with single stripe on pv0

When trying to get LogicalVolumes, and trying to parse LV2, discutils accidentally tries to read it from /dev/sda2 since it searchs all VolumeGroups for pv0 which actually exists differently for both VG1 and VG2.

The solution, to search only within the physical volume list of the VolumeGroup that LogicalVolume is part of.

Branch: [lvm_multi_vpg_same_pv]